1.完成增删改查的基本用法. 2.接入了不少平常使用的一些控件.

+++ app/src/main/java/cn/chenzhongjin/greendao/sample/database/UserDao.java
@@ -26,7 +26,7 @@ public class UserDao extends AbstractDao<User, Long> {
         public final static Property Id = new Property(0, Long.class, "id", true, "_id");
         public final static Property Name = new Property(1, String.class, "name", false, "NAME");
         public final static Property Sex = new Property(2, String.class, "sex", false, "SEX");
-        public final static Property PhoneNumber = new Property(3, Integer.class, "phoneNumber", false, "PHONE_NUMBER");
+        public final static Property PhoneNumber = new Property(3, Long.class, "phoneNumber", false, "PHONE_NUMBER");
         public final static Property UpdateTime = new Property(4, Long.class, "updateTime", false, "UPDATE_TIME");
     };
 
@@ -79,7 +79,7 @@ public class UserDao extends AbstractDao<User, Long> {
             stmt.bindString(3, sex);
         }
  
-        Integer phoneNumber = entity.getPhoneNumber();
+        Long phoneNumber = entity.getPhoneNumber();
         if (phoneNumber != null) {
             stmt.bindLong(4, phoneNumber);
         }
@@ -109,7 +109,7 @@ public class UserDao extends AbstractDao<User, Long> {
             cursor.isNull(offset + 0) ? null : cursor.getLong(offset + 0), // id
             cursor.isNull(offset + 1) ? null : cursor.getString(offset + 1), // name
             cursor.isNull(offset + 2) ? null : cursor.getString(offset + 2), // sex
-            cursor.isNull(offset + 3) ? null : cursor.getInt(offset + 3), // phoneNumber
+            cursor.isNull(offset + 3) ? null : cursor.getLong(offset + 3), // phoneNumber
             cursor.isNull(offset + 4) ? null : cursor.getLong(offset + 4) // updateTime
         );
         return entity;
@@ -121,7 +121,7 @@ public class UserDao extends AbstractDao<User, Long> {
         entity.setId(cursor.isNull(offset + 0) ? null : cursor.getLong(offset + 0));
         entity.setName(cursor.isNull(offset + 1) ? null : cursor.getString(offset + 1));
         entity.setSex(cursor.isNull(offset + 2) ? null : cursor.getString(offset + 2));
-        entity.setPhoneNumber(cursor.isNull(offset + 3) ? null : cursor.getInt(offset + 3));
+        entity.setPhoneNumber(cursor.isNull(offset + 3) ? null : cursor.getLong(offset + 3));
         entity.setUpdateTime(cursor.isNull(offset + 4) ? null : cursor.getLong(offset + 4));
      }
